requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.6
django-extensions==3.2.1
django-filter==22.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==16.6.1
idna==3.4
inflection==0.5.1
jwcrypto==1.4.2
Markdown==3.4.1
mysqlclient==2.1.1
oauthlib==3.2.2
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
toposort==1.9
tzdata==2022.7
urllib3==1.26.14
utils==1.0.1
wrapt==1.14.1
